+++ src/Components/Quiz.js
@@ -85,10 +85,6 @@ const Quiz = ({questions, category, name, difficulty, amount, datas, setDatas}) 
         setShowScore(true);
     }
 
-    const [currentSecs, setCurrentSecs] = useState();
-    const [currentMins, setCurrentMins] = useState();
-    const [currentHours, setCurrentHours] = useState();
-
     // const currentSecs = new Date().getSeconds();
     // const currentMins = new Date().getMinutes();
     // const currentHours = new Date().getHours();
@@ -97,7 +93,6 @@ const Quiz = ({questions, category, name, difficulty, amount, datas, setDatas}) 
         // setCurrentSecs(new Date().getSeconds());
         // setCurrentMins(new Date().getMinutes());
         // setCurrentHours(new Date().getHours());
-        const items = [...datas];
         let data = {
             "min": new Date().getTime(),
             "score": score,
@@ -106,37 +101,57 @@ const Quiz = ({questions, category, name, difficulty, amount, datas, setDatas}) 
             "amount": questions.length, 
             "date": new Date().toLocaleString(),
         }
-        items.push(data);
-        setDatas(items);
+        setDatas((old) => [...old, data]);
     }
-
-    // console.log(questions)
-
-    // console.log(datas)
-    // console.log(questions.category)
-
-    // console.log(questions.length)
-    // if (questions.length <= 0) {
-    //     console.log("tidak ada")
-    // }
+    console.log(datas)
 
     const centerElement = {
         left: "50%",
         transform: "translate(-50%, -20%)",
     }
 
+    const height = {
+        height: '85vh'
+    }
+
+    useEffect(() => {
+        // if (currentQuiz >= questions.length) {
+        //     pushData();
+        // } else if (currentQuiz === 1 || questions.length === 1) {
+        //     if (showScore) {
+        //         pushData();
+        //         console.log('baru');
+        //     }
+        // }
+        if (showScore) {
+            console.log('lol bojo')
+        }
+        if (currentQuiz === 0 || questions.length === 1) {
+            if (currentQuiz >= questions.length) {
+                pushData();
+            }
+        } else {
+            if (questions.length !== 1) {
+                if (currentQuiz >= questions.length) {
+                    pushData();
+                }
+            }
+        }
+    }, [currentQuiz, questions])
+
+    console.log(score)
 
     return(
-        <div className='bg-black h-screen'>
-            {showText && (<h2 className='text-white font-bold text-3xl pt-12'>Welcome!! <span className='text-lime-400'>{name}</span></h2>)}
-            <div className='top-1/2 absolute h-screen ' style={centerElement}>
+        <div className='bg-black h-screen p-3'>
+            {showText && (<h2 className='text-white font-bold text-3xl mt-14'>Welcome!! <span className='text-lime-400'>{name}</span></h2>)}
+            <div className='flex justify-center items-center' style={height}>
             {questions.length > 0 ? (
                 (currentQuiz >= questions.length ? (
                     <div className='bg-lime-400 w-fit p-6 rounded-xl'>    
                             <h2>Your score is {score} out of {questions.length}</h2>
                             <div className='flex flex-col'>
                             <button onClick={() => {setCurrentQuiz(0); setScore(0)}} className="bg-black text-white p-3 rounded-xl m-3">Try Again</button>
-                                <button onClick={pushData}>
+                                <button>
                                 <Link to='/form'>Go Back</Link>
                                 </button>
                             </div>
@@ -144,11 +159,18 @@ const Quiz = ({questions, category, name, difficulty, amount, datas, setDatas}) 
                             
                     ) : (
                     <div>
-                        <h2 className='text-white font-bold text-3xl mb-3'><span className='text-lime-300'>Category:</span> {questions[currentQuiz].category}</h2>
+                        <div className='flex justify-between items-center my-3'>
+                            <h2 className='text-white font-bold text-3xl mb-3'><span className='text-lime-300'>Category:</span> {questions[currentQuiz].category}</h2>
+                            <button className='bg-red-500 text-white p-3 rounded-xl '><Link to='/form'>Cancel</Link></button>
+                        </div>
                         <div className='bg-lime-400 w-fit p-6 rounded-xl'>
                         
                             <div className='w-fit'>
-                                <h2 className='white'>{minutes} : {seconds}</h2>
+                                <div className='flex justify-between mb-4'>
+                                    <div></div>
+                                    <h2 className='white font-bold bg-white px-3 py-2.5 rounded-lg flex'>
+                                    0{minutes} : {seconds}</h2>
+                                </div>
                                 <div className='flex justify-between items-center'>
                                     <div>
                                         <h2 className='mr-2'>{questions[currentQuiz]?.question}</h2>
@@ -180,7 +202,7 @@ const Quiz = ({questions, category, name, difficulty, amount, datas, setDatas}) 
                 <div className='text-center'>
                     <h2 className='text-white'>Sorry, No Result</h2>
                     <button className="bg-lime-400 p-3 rounded-xl m-3 font-semibold">
-                        <Link to='/'>Try Another Question</Link>
+                        <Link to='/form'>Try Another Question</Link>
                     </button>
                 </div>
             )}
